requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
openai>=1.12.0
flask>=3.0.2
pytz>=2024.1
requests>=2.28.1